Fix major bug in renvid() and renlast().
The bug would replace the general VidHop metadata dir variable, to a directory with the title of the last video.

The result of that was, that all following downloads would place their metadata in /vidhop/metadata/<title_of_last_video>.

This bug was pretty bad and I'll have to fix my own VidHop data now.

If you never use renvid() or renlast(), there's nothing to worry about.